+++ robot_hat/mock/__init__.py
@@ -1,0 +1,3 @@
+from robot_hat.mock.uart import MockUART
+
+__all__ = ["MockUART"]